FINAL VERSION (INCOMPLETE)
This project will remain incomplete, first because I have an extreme dislike
for C# which is antagonistic to efficient computational models (such as the
scope of this project requires), and second because the very purpose of
building this project, in anything in this language, is about to expire.

This code does not currently work.  It is missing several States in it's HTTP
Parser as can be seen in the compiler errors.  And when I do finaly finish this
work, it is uncertain if it will be in C/C++ or a custom derived language
considering the ongoing polution of those languages with high-level features
requireing undue overhead making them far less efficient than they are turned
to for being.  (Count RISC steps, not seconds.)